+++ src/js/sidenav.js
@@ -24,7 +24,6 @@ class Sidenav {
     this.sidenavTriggers = document.querySelectorAll('.sidenav-trigger');
     this.isActive = false;
     this.isFixed = this.el.classList.contains('fixed');
-    this.isInLayout = document.body.classList.contains('fixed-sidenav');
     this.isLarge = this.el.classList.contains('large');
 
     /**
@@ -39,7 +38,7 @@ class Sidenav {
       this._createOverlay();
     }
     this._setup();
-    this._handleSidenavLarge();
+    this.el.classList.contains('large') ? document.body.classList.add('sidenav-large') : '';
   }
 
   /**
@@ -63,25 +62,6 @@ class Sidenav {
     this.overlayElement = document.createElement('div');
     this.overlayElement.classList.add('sidenav-overlay');
     this.overlayElement.dataset.target = this.el.id;
-  }
-
-  /**
-   * Handle Sidenav large when is in layout
-   */
-  _handleSidenavLarge() {
-    if (this.isInLayout && this.isLarge && window.innerWidth >= 960) {
-      const elementWidth = this.el.offsetWidth;
-      const navFixedEl = document.querySelector('.navbar-fixed .navbar');
-      navFixedEl ? (navFixedEl.style.paddingLeft = elementWidth + 8 + 'px') : '';
-      document.querySelector('header').style.paddingLeft = elementWidth + 'px';
-      document.querySelector('main').style.paddingLeft = elementWidth + 'px';
-      document.querySelector('footer').style.paddingLeft = elementWidth + 8 + 'px';
-
-      if (document.body.classList.contains('under-navbar')) {
-        document.querySelector('header').style.paddingLeft = '';
-        navFixedEl ? (navFixedEl.style.paddingLeft = 8 + 'px') : '';
-      }
-    }
   }
 
   /**